Move TooltipAdder outside of the adders package
