+++ src/main/java/de/hysky/skyblocker/skyblock/item/tooltip/TooltipAdder.java
@@ -1,7 +1,6 @@
-package de.hysky.skyblocker.skyblock.item.tooltip.adders;
+package de.hysky.skyblocker.skyblock.item.tooltip;
 
 import de.hysky.skyblocker.skyblock.ChestValue;
-import de.hysky.skyblocker.skyblock.item.tooltip.TooltipManager;
 import net.minecraft.screen.slot.Slot;
 import net.minecraft.text.Text;
